+++ src/ExAs/Assertions/MemberAssertions/Enumerables/AnyFulfilsPredicateAssertion.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using ExAs.Results;
 using ExAs.Utils;
@@ -8,11 +9,11 @@ using ExAs.Utils.SystemExtensions;
 
 namespace ExAs.Assertions.MemberAssertions.Enumerables
 {
-    public class NoneFulfilsPredicateAssertion<T> : IAssertValue<IEnumerable<T>>
+    public class AnyFulfilsPredicateAssertion<T> : IAssertValue<IEnumerable<T>>
     {
         private readonly Expression<Func<T, bool>> predicate;
 
-        public NoneFulfilsPredicateAssertion(Expression<Func<T, bool>> predicate)
+        public AnyFulfilsPredicateAssertion(Expression<Func<T, bool>> predicate)
         {
             this.predicate = predicate;
         }
@@ -20,11 +21,10 @@ namespace ExAs.Assertions.MemberAssertions.Enumerables
         public ValueAssertionResult AssertValue(IEnumerable<T> actual)
         {
             var actualList = actual.ToReadOnly();
-            bool succeeded = predicate != null && !actualList.Any_NullAware(predicate);
             return new ValueAssertionResult(
-                succeeded,
+                actualList.Any_NullAware(predicate),
                 actualList.ToValueString(),
-                ComposeLog.Expected($"none fulfils {predicate.ToValueString()}"));
+                ComposeLog.Expected($"any fulfils {predicate.ToValueString()}"));
         }
     }
 }